requirements.txt
PyMuPDF==1.23.5
python-docx==0.8.11
google-generativeai==0.3.2
spacy==3.7.2
streamlit==1.28.0
python-dotenv==1.0.0
pandas==2.0.3